Remove affine symbolic expression.
