+++ executor/src/witgen/jit/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod affine_symbolic_expression;
 mod algebraic_to_quadratic;
 mod block_machine_processor;
 pub(crate) mod code_cleaner;
@@ -10,7 +9,7 @@ mod identity_queue;
 mod interpreter;
 mod processor;
 mod prover_function_heuristics;
-pub(crate) mod quadratic_symbolic_expression;
+mod quadratic_symbolic_expression;
 mod single_step_processor;
 mod symbolic_expression;
 mod variable;
